+++ docs/changelog.rst
@@ -7,6 +7,8 @@ Changelog
 * :feature:`4487` Users can now customize the order of how addresses are resolved to human readable names.
 * :feature:`5001` The PnL report can now be generated with the Highest-In First Out (HIFO) accounting method.
 * :feature:`1793` The PnL report can now be generated with the average cost basis accounting method.
+* :bug:`-` Transfers between tracked accounts will now have a correct label in the UI.
+* :bug:`5038` Premium users with big databases should no longer see the error: "Upload data to server died with exception: database plaintext is locked"
 
 * :release:`1.26.2 <2022-12-09>`
 * :bug:`5187` Fix issue where the blockchain balances were refreshed, when only changing the label or the tags of an account.